feat(ui): add prop types to old component
