+++ libs/ui/src/components/checkbox/Checkbox.data.ts
@@ -1,34 +1,16 @@
 import { ReactNodeI } from '@codelab/graph'
-import { PropsFromKeys } from '@codelab/props'
-import { TextProps } from '../text/Text'
-
-export const checkboxPropKeys = [
-  'autoFocus',
-  'checked',
-  'defaultChecked',
-  'disabled',
-  'indeterminate',
-  'onChange',
-] as const
-
-export const checkboxGroupPropKeys = [
-  'defaultValue',
-  'disabled',
-  'name',
-  'options',
-  'value',
-  'onChange',
-] as const
-
-type CheckBoxProps = PropsFromKeys<typeof checkboxPropKeys[number]>
-
-type CheckboxGroupProps = PropsFromKeys<typeof checkboxGroupPropKeys[number]>
+import { TextProps } from '../text/Text.data'
+import { CheckboxGroupProps, CheckBoxProps } from './Checkbox.types'
 
 export const checkboxData: ReactNodeI<
   CheckBoxProps | CheckboxGroupProps | TextProps
 > = {
   type: 'Checkbox',
   nodeType: 'React',
+  props: {
+    autoFocus: false,
+    value: 'Checkbox',
+  },
   children: [
     {
       type: 'Text',
