+++ libs/ui/src/components/dropdown/Dropdown.data.ts
@@ -1,37 +1,6 @@
 import { ReactNodeI } from '@codelab/graph'
-import { PropsFromKeys } from '@codelab/props'
-import { TextProps } from '../text/Text'
-
-export const dropdownPropKeys = [
-  'arrow',
-  'disabled',
-  'getPopupContainer',
-  'overlay',
-  'overlayClassName',
-  'overlayStyle',
-  'placement',
-  'trigger',
-  'visible',
-  'onVisibleChange',
-] as const
-
-export const dropdownButtonPropKeys = [
-  'disabled',
-  'icon',
-  'overlay',
-  'placement',
-  'size',
-  'trigger',
-  'type',
-  'visible',
-  'onClick',
-  'onVisibleChange',
-  'buttonsRender',
-] as const
-
-type DropdownProps = PropsFromKeys<typeof dropdownPropKeys[number]>
-
-type DropdownButtonProps = PropsFromKeys<typeof dropdownButtonPropKeys[number]>
+import { TextProps } from '../text/Text.data'
+import { DropdownButtonProps, DropdownProps } from './Dropdown.types'
 
 export const dropdownData: ReactNodeI<
   DropdownProps | TextProps | DropdownButtonProps
