+++ libs/ui/src/components/text/Text.data.tsx
@@ -1,12 +1,13 @@
-import { InputNode } from '@codelab/graph'
+import { ReactNodeI } from '@codelab/graph'
 import { PropsFromKeys } from '@codelab/props'
 
 export const textPropKeys = ['value'] as const
 
 export type TextProps = PropsFromKeys<typeof textPropKeys[number]>
 
-export const textData: InputNode<TextProps> = {
+export const textData: ReactNodeI<TextProps> = {
   type: 'Text',
+  nodeType: 'React',
   props: {
     value: 'Some text',
   },
